% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_contributions_glmnet.R
\name{contributionsServer}
\alias{contributionsServer}
\title{Contributions Module Server}
\usage{
contributionsServer(id, model_module, data_module)
}
\arguments{
\item{id}{Module namespace ID.}

\item{model_module}{Reactive list from \code{\link[=modelingServer]{modelingServer()}}.}

\item{data_module}{Reactive list from \code{\link[=dataImportServer]{dataImportServer()}}.}
}
\value{
No return value, called for side effects (renders UI outputs).
}
\description{
Contributions Module Server
}
