% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_contributions_glmnet.R
\name{contributionsUI}
\alias{contributionsUI}
\title{Contributions Module UI}
\usage{
contributionsUI(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\value{
A Shiny tagList.
}
\description{
Contributions Module UI
}
