% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_correlation_glmnet.R
\name{correlationServer}
\alias{correlationServer}
\title{Correlation Matrix Module Server}
\usage{
correlationServer(id, data_module)
}
\arguments{
\item{id}{Module namespace ID.}

\item{data_module}{Reactive list from \code{\link[=dataImportServer]{dataImportServer()}}.}
}
\value{
No return value, called for side effects (renders UI outputs).
}
\description{
Correlation Matrix Module Server
}
