% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_diagnostics_glmnet.R
\name{diagnosticsServer}
\alias{diagnosticsServer}
\title{Diagnostics and Plots Module Server}
\usage{
diagnosticsServer(id, model_module)
}
\arguments{
\item{id}{Module namespace ID.}

\item{model_module}{Reactive list returned by \code{\link[=modelingServer]{modelingServer()}}.}
}
\value{
\code{NULL} (side effects only: renders plots and downloads).
}
\description{
Server logic for diagnostic plots using ggplot2.
}
\examples{
if (interactive()) {
  server <- function(input, output, session) {
    data_out <- dataImportServer("data", shiny::reactiveVal("general"))
    model_out <- modelingServer("model", data_out)
    diagnosticsServer("diag", model_out)
  }
}
}
