% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_earth_import_glmnet.R
\name{earthImportServer}
\alias{earthImportServer}
\title{Earth Import Module -- Server}
\usage{
earthImportServer(id)
}
\arguments{
\item{id}{Shiny module namespace ID.}
}
\value{
A list with components:
\describe{
\item{earth_import}{Reactive containing a \code{glmnetUI_earth_import}
object, or \code{NULL}.}
\item{reset}{Function to clear the imported earth data.}
}
}
\description{
Earth Import Module -- Server
}
