% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthui-package.R
\docType{package}
\name{earthUI-package}
\alias{earthUI}
\alias{earthUI-package}
\title{earthUI: Interactive 'shiny' GUI for the 'earth' Package}
\description{
Provides a 'shiny'-based graphical user interface for the 'earth' package, enabling interactive building and exploration of Multivariate Adaptive Regression Splines (MARS) models. Features include data import from CSV and 'Excel' files, automatic detection of categorical variables, interactive control of interaction terms via an allowed matrix, comprehensive model diagnostics with variable importance and partial dependence plots, and publication-quality report generation via 'Quarto'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wcraytor/earthUI}
  \item Report bugs at \url{https://github.com/wcraytor/earthUI/issues}
}

}
\author{
\strong{Maintainer}: William Craytor \email{bcraytor@proton.me}

}
\keyword{internal}
