% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_equation_glmnet.R
\name{equationServer}
\alias{equationServer}
\title{Model Equation Module Server}
\usage{
equationServer(id, model_module, data_module)
}
\arguments{
\item{id}{Module namespace ID.}

\item{model_module}{Reactive list from \code{\link[=modelingServer]{modelingServer()}}.}

\item{data_module}{Reactive list from \code{\link[=dataImportServer]{dataImportServer()}}.}
}
\value{
A reactive list containing the equation LaTeX strings.
}
\description{
Model Equation Module Server
}
