% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_equation_glmnet.R
\name{equationUI}
\alias{equationUI}
\title{Model Equation Module UI}
\usage{
equationUI(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\value{
A Shiny tagList.
}
\description{
Model Equation Module UI
}
