% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_results_mgcv.R
\name{glance_gam}
\alias{glance_gam}
\title{Glance at GAM Model Fit}
\usage{
glance_gam(gam_result)
}
\arguments{
\item{gam_result}{An \code{mgcvUI_result} from \code{\link[=fit_gam]{fit_gam()}}.}
}
\value{
A one-row data frame (tibble).
}
\description{
Returns a one-row data frame of model-level statistics using
\code{\link[broom:reexports]{broom::glance()}}.
}
\examples{
specs <- list(list(vars = "wt", type = "s", bs = "tp", k = NULL))
res <- fit_gam(mtcars, "mpg", specs)
glance_gam(res)
}
