% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_importance_glmnet.R
\name{importanceServer}
\alias{importanceServer}
\title{Variable Importance Module Server}
\usage{
importanceServer(id, model_module, data_module)
}
\arguments{
\item{id}{Module namespace ID.}

\item{model_module}{Reactive list from \code{\link[=modelingServer]{modelingServer()}}.}

\item{data_module}{Reactive list from \code{\link[=dataImportServer]{dataImportServer()}}.}
}
\value{
No return value, called for side effects (renders UI outputs).
}
\description{
Variable Importance Module Server
}
