% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_importance_glmnet.R
\name{importanceUI}
\alias{importanceUI}
\title{Variable Importance Module UI}
\usage{
importanceUI(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\value{
A Shiny tagList.
}
\description{
Variable Importance Module UI
}
