% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_model_mgcv.R
\name{mod_model_results_ui}
\alias{mod_model_results_ui}
\title{Model Results -- UI (main panel)}
\usage{
mod_model_results_ui(id)
}
\arguments{
\item{id}{Shiny module namespace ID.}
}
\value{
A \link[shiny:reexports]{shiny::tagList}.
}
\description{
Tabbed output panel for model results, plots, and diagnostics.
}
