% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results_mgcv.R
\name{plot_by_smooth_single}
\alias{plot_by_smooth_single}
\title{Static Factor-by-Smooth Contribution Plot}
\usage{
plot_by_smooth_single(gam_result, variable, by_var)
}
\arguments{
\item{gam_result}{An \code{mgcvUI_result} from \code{\link[=fit_gam]{fit_gam()}}.}

\item{variable}{Character name of the continuous variable.}

\item{by_var}{Character name of the factor variable.}
}
\value{
A ggplot object.
}
\description{
Creates a ggplot line plot for a factor-by-smooth term,
showing one curve per factor level.
}
