% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regproj.R
\name{regproj_last_file}
\alias{regproj_last_file}
\alias{regproj_last_file_path}
\alias{regproj_last_file_get}
\alias{regproj_last_file_set}
\title{Last-used input file marker (per project)}
\usage{
regproj_last_file_path(project_path, os = os_detect())

regproj_last_file_get(project_path, os = os_detect())

regproj_last_file_set(project_path, basename, os = os_detect())
}
\arguments{
\item{project_path}{Absolute project path.}

\item{os}{OS segment. Defaults to \code{\link[=os_detect]{os_detect()}}.}

\item{basename}{File basename to remember.}
}
\value{
Path / basename / invisible path.
}
\description{
Stores the basename of the most-recently-selected input file in
\verb{<project>/<os>/.regproj-last}, so it can be auto-selected the next
time the project is opened.
}
