% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_summary_glmnet.R
\name{summaryServer}
\alias{summaryServer}
\title{Model Summary Module Server}
\usage{
summaryServer(id, model_module, data_module, purpose)
}
\arguments{
\item{id}{Module namespace ID.}

\item{model_module}{Reactive list from \code{\link[=modelingServer]{modelingServer()}}.}

\item{data_module}{Reactive list from \code{\link[=dataImportServer]{dataImportServer()}}.}

\item{purpose}{Reactive returning current purpose mode.}
}
\value{
A reactive list with summary statistics.
}
\description{
Model Summary Module Server
}
