% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_summary_glmnet.R
\name{summaryUI}
\alias{summaryUI}
\title{Model Summary Module UI}
\usage{
summaryUI(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\value{
A Shiny tagList.
}
\description{
Model Summary Module UI
}
