% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_results_mgcv.R
\name{tidy_gam}
\alias{tidy_gam}
\title{Tidy GAM Results with broom}
\usage{
tidy_gam(gam_result)
}
\arguments{
\item{gam_result}{An \code{mgcvUI_result} from \code{\link[=fit_gam]{fit_gam()}}.}
}
\value{
A data frame (tibble) of tidied model terms.
}
\description{
Returns a tidy data frame of model terms using \code{\link[broom:reexports]{broom::tidy()}}.
}
\examples{
specs <- list(list(vars = "wt", type = "s", bs = "tp", k = NULL))
res <- fit_gam(mtcars, "mpg", specs)
tidy_gam(res)
}
