% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_partial_dependence}
\alias{plot_partial_dependence}
\title{Plot partial dependence}
\usage{
plot_partial_dependence(earth_result, variable, n_grid = 50L)
}
\arguments{
\item{earth_result}{An object of class \code{"earthui_result"} as returned by
\code{\link[=fit_earth]{fit_earth()}}.}

\item{variable}{Character string. Name of the predictor variable to plot.}

\item{n_grid}{Integer. Number of grid points for the partial dependence
calculation. Default is 50.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Creates a partial dependence plot for a selected variable from a fitted
earth model.
}
\examples{
result <- fit_earth(mtcars, "mpg", c("cyl", "disp", "hp", "wt"))
plot_partial_dependence(result, "wt")
}
