% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_qq}
\alias{plot_qq}
\title{Plot Q-Q plot of residuals}
\usage{
plot_qq(earth_result)
}
\arguments{
\item{earth_result}{An object of class \code{"earthui_result"} as returned by
\code{\link[=fit_earth]{fit_earth()}}.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Creates a normal Q-Q plot of the model residuals.
}
