% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_results.R
\name{validate_earthui_result}
\alias{validate_earthui_result}
\title{Validate earthui_result object}
\usage{
validate_earthui_result(x)
}
\arguments{
\item{x}{Object to validate.}
}
\value{
Invisible NULL. Raises error if invalid.
}
\description{
Validate earthui_result object
}
\keyword{internal}
